requirements.txt
streamlit
pandas
python-slugify
bcrypt
